Refactored code. Move audio related code to `src/core/audio.js/
